[#175883367] substitute bPayCard with the one already exist
